redesigned the directory structure
