+++ server/server.js
@@ -1,0 +1,8 @@
+const http = require('http');
+const app = require('./app');
+const port = process.env.PORT || 3000;
+const server = http.createServer(app);
+
+server.listen(port,()=>{
+    console.log(`listning on port ${port}`)
+});
